driver: support external ips
Added support for attaching external IPs to the instance. Both ephemeral
IPs and floating IPs are supported by providing a `TYPE,IP_POOL_ID` pair
to the `OXIDE_EXTERNAL_IPS` flag. The `TYPE` is either `ephemeral` or
`floating` and the `IP_POOL_ID` is a valid UUID for an existing IP pool.