driver: support external ips (#33)
Added support for attaching external IPs to the instance. Both ephemeral
IPs and floating IPs are supported by providing a `TYPE,NAME_OR_ID`
pair to the `oxide-external-ip` flag. The `TYPE` is either `ephemeral`
or `floating` and the `NAME_OR_ID` is the name or ID for an existing IP
pool or floating IP.