driver: add flag to specify network interface ip
Added a new `oxide-nic-ip` string flag to configure the IP address for
the instance's network interface. Perhaps there's an opportunity here to
create flags that allow users to specify multiple network interfaces, but
that can be a follow-up item if there are use cases for that.

This change was introduced because there were cases where an instance's
network interface would not be assigned an IP address unless the IP
address was explicitly specified even though the underlying subnet had
available IP addresses for use.